+++ docker-compose.yml
@@ -12,5 +12,8 @@ services:
     environment:
     - NGINX_HOST=foobar.com
     - NGINX_PORT=80
+    
   proxy:
     image: dockerbar/image-previewer:v0.0.1
+    ports:
+    - "8081:8081"
